pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strobes-shell-agent"
version = "0.1.0"
description = "Strobes Shell Bridge Agent — connects your machine to Strobes for remote command execution"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
dependencies = [
    "websockets>=12.0",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pyinstaller>=6.0",
]

[project.scripts]
strobes-shell-agent = "strobes_shell_agent.cli:main"

[tool.setuptools.packages.find]
include = ["strobes_shell_agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
